refactor: split ChatViewModel into focused sub-objects
