+++ clients/shared/Features/Chat/ChatErrorManager.swift
@@ -1,0 +1,85 @@
+import Foundation
+import os
+#if os(macOS)
+import AppKit
+#elseif os(iOS)
+import UIKit
+#else
+#error("Unsupported platform")
+#endif
+import Network
+
+private let log = Logger(subsystem: Bundle.main.bundleIdentifier ?? "com.vellum.vellum-assistant", category: "ChatErrorManager")
+
+/// Owns error-related @Published properties that were previously part of ChatViewModel.
+/// ChatViewModel holds a reference to this object and forwards reads/writes via
+/// computed properties so every existing call site continues to compile without
+/// modification.
+@MainActor
+public final class ChatErrorManager: ObservableObject {
+
+    /// Human-readable error string shown in the error banner.
+    @Published public var errorText: String?
+
+    /// Typed session error, richer than `errorText` and used for structured retry UI.
+    @Published public var sessionError: SessionError?
+
+    /// Supplemental diagnostic hint shown alongside a daemon connection error.
+    /// Nil when no connection error is active or the error has been dismissed.
+    @Published public var connectionDiagnosticHint: String? = nil
+
+    // MARK: - Connection diagnostics
+
+    /// Map a raw connection error to a short, actionable diagnostic hint.
+    ///
+    /// The hint is shown below the generic "Failed to connect" error banner so
+    /// users can understand *why* the connection failed without opening the
+    /// debug panel or contacting support.
+    public static func connectionDiagnosticHint(for error: Error) -> String? {
+        #if os(macOS)
+        if let nwErr = error as? NWError {
+            switch nwErr {
+            case .posix(let code):
+                switch code {
+                case .ECONNREFUSED:
+                    return "Daemon is not accepting connections — it may still be starting."
+                case .ENOENT:
+                    return "Daemon socket not found — the daemon may not be running."
+                case .ETIMEDOUT:
+                    return "Connection timed out — the daemon may be unresponsive or the socket path may be wrong."
+                default:
+                    return "POSIX error \(code.rawValue): \(nwErr.localizedDescription)"
+                }
+            default:
+                return nwErr.localizedDescription
+            }
+        }
+        if let authErr = error as? DaemonClient.AuthError {
+            switch authErr {
+            case .missingToken:
+                return "Session token not found — try restarting the daemon."
+            case .timeout:
+                return "Authentication timed out — daemon may be overloaded."
+            case .rejected:
+                return "Session token rejected — token may have changed; try restarting the daemon."
+            }
+        }
+        #endif
+        // HTTP transport / iOS
+        if let urlErr = error as? URLError {
+            switch urlErr.code {
+            case .notConnectedToInternet:
+                return "Device is not connected to the internet."
+            case .timedOut:
+                return "Gateway request timed out — check your host and port."
+            case .cannotConnectToHost:
+                return "Cannot connect to gateway host — verify the address and that the daemon is running."
+            case .networkConnectionLost:
+                return "Network connection lost."
+            default:
+                return urlErr.localizedDescription
+            }
+        }
+        return nil
+    }
+}
